chore: re-organize folder to follow conventions
