+++ packages/components/bolt-button/index.js
@@ -1,0 +1,28 @@
+import { polyfillLoader } from '@bolt/core/polyfills';
+
+polyfillLoader.then(res => {
+  import(/* webpackMode: 'eager', webpackChunkName: 'bolt-button' // */ './src/button');
+});
+
+
+
+
+
+// import { polyfillLoader } from '@bolt/core/polyfills';
+//
+// polyfillLoader.then(res => {
+//   if (!window.customElements.get('replace-with-children')) {
+//     import(/*
+//       webpackMode: 'eager',
+//       webpackChunkName: 'replace-with-children'
+//     */ '@bolt/core/elements/replace-with-children');
+//   }
+//
+//   if (!window.customElements.get('bolt-button')) {
+//     import(/*
+//       webpackMode: 'eager',
+//       webpackChunkName: 'bolt-button'
+//     */ './src/button');
+//   }
+// });
+//
