+++ packages/components/bolt-button/index.scss
@@ -1,0 +1,1 @@
+@import 'src/button.scss';
