+++ packages/components/bolt-button/package.json
@@ -1,22 +1,33 @@
 {
   "name": "@bolt/components-button",
-  "style": "src/button.scss",
-  "main": "src/button.js",
-  "description": "Button Component",
+  "description": "Button Component in Bolt",
   "keywords": [
-    "eyeglass-module",
     "bolt design system",
     "bolt",
-    "ui toolkit",
-    "design system"
+    "css framework",
+    "design system",
+    "components"
   ],
   "version": "2.3.0",
   "sourceFilesFrom": "0.9.8",
-  "author": "Salem Ghoweri",
+  "maintainers": [
+    {
+      "name": "Salem Ghoweri",
+      "email": "me@salemghoweri.com",
+      "web": "https://github.com/sghoweri"
+    },
+    {
+      "name": "Mike Mai",
+      "email": "boss@mikemai.net",
+      "web": "http://mikemai.net/"
+    }
+  ],
   "homepage": "https://boltdesignsystem.com",
   "license": "MIT",
   "repository": "https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-button",
   "bugs": "https://github.com/bolt-design-system/bolt/issues",
+  "style": "index.scss",
+  "main": "index.js",
   "publishConfig": {
     "access": "public"
   },
