+++ packages/components/bolt-button/src/_button-settings-and-tools.scss
@@ -1,3 +1,32 @@
+/* ------------------------------------ *\
+   Button settings and tools
+\* ------------------------------------ */
+
+/**
+ * Settings
+ */
+
+$bolt-color-button-text: bolt-color(white);
+$bolt-color-primary: bolt-color(yellow);
+$bolt-color-secondary: bolt-color(indigo, light);
+$bolt-button-shadow: $bolt-shadow--small;
+$bolt-button-shadow--hover: $bolt-shadow--large;
+$bolt-button-border-width: $bolt-border-width;
+$bolt-button-border-style: $bolt-border-style;
+$bolt-button-border-radii: (
+  'regular': $bolt-border-radius,
+  'full': 50em,
+);
+$bolt-button-transition: $bolt-transition;
+$bolt-button-translate--hover: translate3d(0, -1px, 0);
+$bolt-button-translate--active: translate3d(0, 1px, 0);
+
+
+
+/**
+ * Tools
+ */
+
 @function bolt-button-border-radius($name) {
   @return map-get($bolt-button-border-radii, $name);
 }
@@ -61,9 +90,7 @@
   }
 }
 
-/**
-  * Button Colors / Types
-  */
+// Button Colors / Types
 @mixin bolt-button-style--primary() {
   color: bolt-theme(text-on-primary);
   border-color: bolt-theme(primary);
